+++ transaction/__init__.py
@@ -11,13 +11,17 @@
 # FOR A PARTICULAR PURPOSE.
 #
 ############################################################################
-"""Exported transaction functions.
+"""``transaction`` module: Exported transaction functions.
 
-$Id$
 """
 
+#: Default implementation of `~ITransaction`
 from transaction._transaction import Transaction
+#: Default implementation of `~ISavepoint`
+from transaction._transaction import Savepoint
+#: A single-threaded `~ITransactionManager`
 from transaction._manager import TransactionManager
+#: A thread-safe `~ITransactionManager`
 from transaction._manager import ThreadTransactionManager
 
 # NB: "with transaction:" does not work under Python 3 because they worked
@@ -25,13 +29,23 @@ from transaction._manager import ThreadTransactionManager
 # via getattr and getattribute; see http://bugs.python.org/issue12022.  On
 # Python 3, you must use ``with transaction.manager`` instead.
 
+#: The default transaction manager (a `~.ThreadTransactionManager`). All other functions in
+#: this module refer to this object.
 manager = ThreadTransactionManager()
+#: See `.ITransactionManager.get`
 get = __enter__ = manager.get
+#: See `.ITransactionManager.begin`
 begin = manager.begin
+#: See `.ITransactionManager.commit`
 commit = manager.commit
+#: See `.ITransactionManager.abort`
 abort = manager.abort
 __exit__ = manager.__exit__
+#: See `.ITransactionManager.doom`
 doom = manager.doom
+#: See `.ITransactionManager.isDoomed`
 isDoomed = manager.isDoomed
+#: See `.ITransactionManager.savepoint`
 savepoint = manager.savepoint
+#: See `.ITransactionManager.attempts`
 attempts = manager.attempts